ClientAdvisor/App/requirements.txt
azure-identity==1.15.0
# Flask[async]==2.3.2
openai==1.6.1
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
quart==0.19.4
uvicorn==0.24.0
aiohttp==3.9.2
gunicorn==20.1.0
quart-session==3.0.0
pyodbc==5.2.0
httpx==0.27.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
flake8==7.1.1
black==24.8.0
autoflake==2.3.1
isort==5.13.2
